pyproject.toml
[project]
name = "feluda"
version = "0.9.4"
description = "A configurable engine for analysing multi-lingual and multi-modal content."
readme = "README.md"
requires-python = ">=3.10"

authors = [{ name = "Tattle Admin", email = "admin@tattle.co.in" }]
maintainers = [{ name = "Tattle Admin", email = "admin@tattle.co.in" }]

license = { file = "LICENSE.md" }
keywords = [
    "multilingual",
    "multimodal",
    "indian-languages",
    "search",
    "deep-learning",
]

classifiers = [
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: Linguistic",
    "Topic :: Multimedia :: Video",
    "Topic :: Multimedia :: Sound/Audio",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = ["dacite>=1.9.2", "pyyaml>=6.0.2"]

# deps that are optional
[project.optional-dependencies]
dev = [
    # for reading, editing pyproject.toml files
    "tomlkit>=0.13.2",
    # code linter and formatter
    "ruff>=0.11.4",
    # runs checks before each git commits
    "pre-commit>=4.2.0",
    # for Jupyter support and interface
    "jupyter>=1.1.1",
    "notebook>=7.3.3",
    "pytest>=8.3.2",
    # documentation dependencies
    "sphinx>=7.0.0",
    "sphinx-rtd-theme>=1.3.0",
    "myst-nb>=1.0.0",
    "sphinx-copybutton>=0.5.0",
    "sphinx-design>=0.4.0",
    # docstring style checking
    "pydocstyle>=6.3.0",
    "docformatter>=0.6.0",
    # benchmark dependencies
    "psutil>=5.9.0",
    "memory-profiler>=0.61.0",
    "numpy>=1.26,<2.2.0",
]
audio = ["pydub>=0.25.1", "wget>=3.2", "boto3>=1.38.42"]

video = ["wget>=3.2", "werkzeug>=3.1.3", "boto3>=1.38.42"]

image = [
    "numpy>=1.26,<2.2.0",
    "pillow>=11.1.0",
    "wget>=3.2",
    "requests>=2.32.3",
    "werkzeug>=3.1.3",
    "boto3>=1.38.42"
]

all = ["feluda[audio,video,image]"]

[project.urls]
Homepage = "https://tattle.co.in/products/feluda/"
Repository = "https://github.com/tattle-made/feluda"
Issues = "https://github.com/tattle-made/feluda/issues"
Changelog = "https://github.com/tattle-made/feluda/blob/main/CHANGELOG.md"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.semantic_release]
version_variable = ["pyproject.toml:project.version"]

[tool.semantic_release.branches.main]
match = "main"
prerelease = false
tag_format = "{name}-{version}"

[tool.uv.workspace]
members = [
    "operators/classify_video_zero_shot",
    "operators/cluster_embeddings",
    "operators/detect_lewd_images",
    "operators/detect_text_in_image",
    "operators/dimension_reduction",
    "operators/image_vec_rep",
    "operators/vid_vec_rep",
    "operators/video_hash",
]

[tool.uv.sources]
feluda = { workspace = true }

[tool.ruff]
target-version = "py310"
line-length = 88
exclude = ["src", "dist", ".docker", ".pytest_cache", ".ruff_cache"]

[tool.hatch.build.targets.wheel]
packages = ["feluda"]
